исправил замечания и добавил сохранение данных активного попапа
