+++ src/view/popup.js
@@ -16,21 +16,6 @@ const CreatePopupElement = (film) => {
   const runtime = getDurationFromMinutes(filmInfo.runtime) || '';
   const genres = filmInfo.genre.split(' ').length > 1 ? 'Gengres' : 'Genre';
 
-  // const {moviename, poster, description, premiereDate, rating, genre, runtime, isInWatchlist, isWatched, actors, isFavorite, pegi, director, writers} = film;
-  // const date = dayjs(premiereDate).format('DD MMM YYYY');
-  // const hours = Math.floor(runtime / 60);
-  // const minutes = Math.floor(runtime) - (hours * 60);
-  // const watchlistClassName = isInWatchlist
-  //   ? 'film-details__control-button--active film-details__control-button--watchlist'
-  //   : 'film-details__control-button--watchlist';
-  // const watchedClassName = isWatched
-  //   ? 'film-details__control-button--active film-details__control-button--watched'
-  //   : 'film-details__control-button--watched';
-  // const favoriteClassName = isFavorite
-  //   ? 'film-details__control-button--active film-details__control-button--favorite'
-  //   : 'film-details__control-button--favorite';
-
-
   return `<section class="film-details" data-id="${id}">
   <form class="film-details__inner" action="" method="get">
     <div class="film-details__top-container">
